gaming-marketing-db-prep/requirements.txt
mysql-connector-python>=8.0.0
faker>=8.0.0
python-dotenv>=0.19.0
